requirements.txt
# CryptoBot Dependencies
#
# Install with: python -m pip install -r requirements.txt

# Core dependencies
pandas>=2.0.0
ccxt>=4.0.0
PyYAML>=6.0

# Data processing
numpy>=1.24.0
ta-lib>=0.4.0  # Technical analysis library (optional, for advanced indicators)

# WebSocket streaming
websockets>=11.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Optional: ML pipeline
scikit-learn>=1.3.0
joblib>=1.3.0
